+++ src/nn/variation/architecture_mutation.py
@@ -10,10 +10,9 @@ from torch.fx.passes.shape_prop import ShapeProp
 from ..individual_graph_module import NeuralNetworkIndividualGraphModule
 from ..variation.utils import (
     node_has_shape, add_specific_node, add_skip_connection, 
-    get_feature_dims, node_has_float_dtype, print_graph_debug_info, get_unique_name
+    get_feature_dims, node_has_float_dtype, print_graph_debug_info, add_branch_nodes
 )
-from ..variation.architecture_adaptation import adapt_node_shape, add_branch_nodes
-from ..visualization import visualize_graph
+from ..variation.architecture_adaptation import adapt_node_shape
 
 
 def mutation_add_linear(individual, **kwargs):
@@ -297,10 +296,22 @@ def _add_node(graph: NeuralNetworkIndividualGraphModule, reference_node: torch.f
         branch2_module = nn.Linear(input_size, branch2_out_size)
         
         # Use the utility function to add branch nodes
-        graph, new_node, skip_output_shape = add_branch_nodes(graph, reference_node, branch1_module, branch2_module)
+        graph, new_node, branch1_node, branch2_node, branch1_shape, branch2_shape = add_branch_nodes(
+            graph, reference_node, branch1_module, branch2_module
+        )
+        
+        # Adapt branch nodes if needed to ensure they have compatible shapes
+        if branch1_shape != branch2_shape:    
+            # Adapt first branch
+            graph, adapted_branch1_node = adapt_node_shape(graph, branch1_node, branch1_shape, branch2_shape, target_user=new_node)
+            # Update the skip connection node's args to use the adapted node
+            new_node.args = (adapted_branch1_node, branch2_node)
+            
+        # Run shape propagation to update metadata
+        ShapeProp(graph).propagate(graph.example_input)
         
         # Get feature dimensions from skip connection output shape
-        new_node_output_shape = get_feature_dims(skip_output_shape)
+        new_node_output_shape = get_feature_dims(new_node.meta['tensor_meta'].shape)
         
         # For branches, input shape is reference node features, output is from skip connection
         new_node_input_shape = ref_feature_shape
