+++ src/nn/variation/utils.py
@@ -7,7 +7,6 @@ import torch.fx
 from torch.fx.passes.shape_prop import ShapeProp
 
 from ..individual_graph_module import NeuralNetworkIndividualGraphModule
-from .architecture_adaptation import ReshapeModule, _adapt_tensor_size, adapt_node_shape_basic, gcf_adapt_node_shape
 
 
 def get_unique_name(graph, base_name: str) -> str:
@@ -131,46 +130,6 @@ def add_skip_connection(graph, second_node, first_node, torch_function=torch.add
 
     return graph, new_node
     
-def adapt_node_shape(graph, node, current_size, target_size, target_user=None, adapt_type='gcf'):
-    """
-    Adapts a node's output shape to match a target size using repetition, adaptive pooling or circular padding.
-    
-    Args:
-        graph: The FX graph
-        node: The node whose shape needs to be adapted
-        current_size: Current size of the node's output, no batch dimension
-        target_size: Desired size of the node's output, no batch dimension
-        target_user: Optional specific node that should use the adapted output. If None, all users will be updated.
-        adapt_type: Type of adaptation to use. Can be 'regular', 'linear', or 'gcf'
-    Returns:
-        graph: The modified graph
-        adapted_node: The node after shape adaptation
-    """
-    # Convert current_size and target_size to tuples if they are not already
-    current_size = tuple(current_size)
-    target_size = tuple(target_size)
-    
-    # If current_size = target_size, return the node
-    if current_size == target_size:
-        return graph, node
-    
-    # Get total elements of current_size and target_size
-    current_total = math.prod(current_size)
-    target_total = math.prod(target_size)
-    
-    # If current_total = target_total, return a reshape node
-    if current_total == target_total:
-        return add_specific_node(
-            graph,
-            node,
-            ReshapeModule(target_size),
-            target_user=target_user
-        )
-    
-    if adapt_type == 'gcf':
-        return gcf_adapt_node_shape(graph, node, current_size, target_size, target_user)
-    else:
-        return adapt_node_shape_basic(graph, node, current_size, target_size, target_user, adapt_type)
     
 def add_branch_nodes(graph: NeuralNetworkIndividualGraphModule, reference_node, branch1_module, branch2_module):
     """
@@ -182,8 +141,13 @@ def add_branch_nodes(graph: NeuralNetworkIndividualGraphModule, reference_node, 
         branch1_module: The module for the first branch
         branch2_module: The module for the second branch
     Returns:
-        graph: The modified graph
-        new_node: The combined node after the branches
+        tuple: (graph, new_node, branch1_node, branch2_node, branch1_shape, branch2_shape)
+            - graph: The modified graph
+            - new_node: The combined node after the branches
+            - branch1_node: The first branch node (may need adaptation)
+            - branch2_node: The second branch node (may need adaptation)
+            - branch1_shape: The shape of branch1_node's output
+            - branch2_shape: The shape of branch2_node's output
     """
     # Generate unique names for the branch nodes
     branch1_name = get_unique_name(graph, "branch1")
@@ -212,27 +176,12 @@ def add_branch_nodes(graph: NeuralNetworkIndividualGraphModule, reference_node, 
     # Infer the shapes of the branch nodes from the metadata, pass through get_feature_dims to remove batch dimension
     branch1_shape = get_feature_dims(branch1_node.meta['tensor_meta'].shape)
     branch2_shape = get_feature_dims(branch2_node.meta['tensor_meta'].shape)
-    
-    # Initialize variables to track the final nodes to use in skip connection
-    final_branch1_node = branch1_node
-    final_branch2_node = branch2_node
-    
-    # Adapt branch nodes if needed to ensure they have compatible shapes
-    if branch1_shape != branch2_shape:    
-        # Adapt first branch
-        graph, final_branch1_node = adapt_node_shape(graph, branch1_node, branch1_shape, branch2_shape)
 
-    # Run shape propagation to update metadata for the branch nodes
-    ShapeProp(graph).propagate(graph.example_input)
-
-    # Get the inferred output of the skip connection from the shape propagation
-    skip_connection_output_shape = final_branch2_node.meta['tensor_meta'].shape
-
-    # Create a skip connection between the two branch nodes using the final adapted nodes
-    with graph.graph.inserting_after(final_branch2_node):
+    # Create a skip connection between the two branch nodes
+    with graph.graph.inserting_after(branch2_node):
         new_node = graph.graph.call_function(
             torch.add,
-            args=(final_branch1_node, final_branch2_node),
+            args=(branch1_node, branch2_node),
         )
     
     # Update connections - replace all uses of reference_node with new_node
@@ -241,7 +190,7 @@ def add_branch_nodes(graph: NeuralNetworkIndividualGraphModule, reference_node, 
     branch1_node.args = (reference_node,)
     branch2_node.args = (reference_node,)
 
-    return graph, new_node, skip_connection_output_shape
+    return graph, new_node, branch1_node, branch2_node, branch1_shape, branch2_shape
 
 def get_feature_dims(shape):
     """
